Progress toward parsing parameters on command line in unit tests.  They are parsing in main() but are not yet available in TEST_CASE() instances.
